requirements.txt
Sphinx
sphinx_rtd_theme
pyyaml